+++ src/modules/app/types/app-service-route.ts
@@ -1,0 +1,3 @@
+export interface IAppServiceRoute {
+  path: string;
+}