+++ src/modules/app/types/app-spec-service.ts
@@ -1,0 +1,11 @@
+import { IAppServiceGithub, IAppServiceRoute } from ".";
+
+export interface IAppSpecService {
+  environment_slug: string;
+  github: IAppServiceGithub;
+  instance_count: number;
+  instance_size_slug: string;
+  name: string;
+  routes: IAppServiceRoute[];
+  run_command: string;
+}