+++ src/modules/app/types/app-spec.ts
@@ -1,0 +1,7 @@
+import { IAppSpecService } from ".";
+
+export interface IAppSpec {
+  name: string;
+  services: IAppSpecService[];
+  region: string;
+}