+++ src/modules/app/index.ts
@@ -1,4 +1,7 @@
 export * from './create-app/create-app';
+export * from './delete-app/delete-app';
 export * from './get-app/get-app';
 export * from './list-apps/list-apps';
+export * from './run-app-detection/run-app-detection';
 export * from './types';
+export * from './update-app/update-app';