+++ src/modules/app/run-app-detection/run-app-detection.ts
@@ -1,0 +1,42 @@
+import { IResponse, IContext } from '../../../types';
+import { IApp, IGithubSpec, IGitSpec } from '..';
+
+export interface IRunAppDetectionComponent {
+  environment_slug: string;
+  run_command: string;
+  strategy: string;
+  types: string[];
+};
+
+export interface IRunAppDetectionApiResponse {
+  components: IRunAppDetectionComponent[];
+}
+
+export interface IRunAppDetectionApiRequest {
+  commit_sha?: string;
+  component_type?: string;
+  git?: GitSpec;
+  github?: GithubSpec;
+}
+
+export type RunAppDetectionResponse = IResponse<IRunAppDetectionApiResponse>;
+
+export const runAppDetection = ({
+  httpClient,
+}: IContext) => ({
+  commit_sha,
+  component_type,
+  git,
+  github,
+}: IRunAppDetectionApiRequest): Promise<Readonly<RunAppDetectionResponse>> => {
+  const path = '/apps/detect';
+  const body = {
+    commit_sha,
+    component_type,
+    git,
+    github,
+  };
+  const url = `${path}`;
+
+  return httpClient.post<IRunAppDetectionApiResponse>(url, body);
+};