+++ src/modules/app/types/app-service-github.ts
@@ -1,4 +1,0 @@
-export interface IAppServiceGithub {
-  branch: string;
-  repo: string;
-}