+++ src/modules/app/types/app-spec-service.ts
@@ -1,8 +1,8 @@
-import { IAppServiceGithub, IAppServiceRoute } from ".";
+import { IGithubSpec, IAppServiceRoute } from ".";
 
 export interface IAppSpecService {
   environment_slug: string;
-  github: IAppServiceGithub;
+  github: IGithubSpec;
   instance_count: number;
   instance_size_slug: string;
   name: string;