+++ src/modules/app/types/git-source-spec.ts
@@ -1,0 +1,4 @@
+export interface IGitSpec {
+  branch: string;
+  repo_clone_url: string;
+}