+++ src/modules/app/types/github-source-spec.ts
@@ -1,0 +1,5 @@
+export interface IGithubSpec {
+  branch: string;
+  deploy_on_push?: boolean;
+  repo: string;
+}